\name{gldist-package}

\alias{gldist-package}

\docType{package}

\title{

  An Asymmetry-Steepness Parameterization of the Generalized
  Lambda Distribution.

}

\description{

  FIXME

}

\author{

  Yohan Chalabi and Diethelm W\"urtz.
  Maintainer: Yohan Chalabi <yohan.chalabi@phys.ethz.ch>

}

\references{

  An Asymmetry-Steepness Parameterization of the Generalized Lambda
  Distribution. Y. Chalabi, D. J. Scott, D. W\"urtz.

}

\keyword{ package }
